Update: Modularise Search component
