+++ src/components/adslotUi/TreePicker/TreePickerNavComponent.jsx
@@ -1,6 +1,6 @@
 import React, { PropTypes } from 'react';
 import Breadcrumb from 'components/alexandria/Breadcrumb/component';
-import Search from 'components/alexandria/SearchComponent';
+import Search from 'components/alexandria/Search/component';
 import SvgSymbol from 'components/alexandria/SvgSymbolComponent';
 import TreePickerPropTypes from 'helpers/propTypes/TreePickerPropTypes';
